% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svymisc.R
\docType{package}
\name{svrepmisc}
\alias{svrepmisc}
\alias{svrepmisc-package}
\title{svrepmisc: Miscellaneous Functions for Replicate Weights}
\description{
Wrapper functions for Complex Surveys using replicate weights.
Takes advantage of \code{\link[survey]{withReplicates}}.
}
