% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svymisc.R
\name{svymultinom}
\alias{svymultinom}
\title{Wrapper for Multinomial Logistic Regression for Replicate Weights}
\usage{
svymultinom(formula, design, subset, ...)
}
\description{
Uses \code{\link[survey]{withReplicates}} and \code{\link[nnet]{multinom}} to generate
coefficients, and standards errors for multinomial logistic regressions
using replicate weights
}
\note{
Output is consistent with SAS's proc surveylogistic's multinomial
survey output
}
\seealso{
\code{\link[survey]{withReplicates}} \code{\link[nnet]{multinom}}
}

