% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svymisc.R
\name{svynb}
\alias{svynb}
\alias{svyrlm}
\title{Wrapper for MASS functions}
\usage{
svynb(formula, design, subset, ..., scale.weights = FALSE)

svyrlm(formula, design, subset, ..., scale.weights = FALSE)
}
\arguments{
\item{formula}{Model formula}

\item{design}{Survey design from \code{\link[survey]{svrepdesign}}}

\item{subset}{Expression to select a subpopulation}

\item{...}{Other arugments passed to \code{\link[MASS]{glm.nb}} or \code{\link[MASS]{rlm}}}

\item{scale.weights}{Indicate whether to rescale weights (defaults to false)}
}
\description{
Wrapper for several functions from \pkg{MASS} for replicate weights
}
\references{
Lumley, Thomas. Complex Surveys: A Guide to Analisys Using R.
 Hoboken, NJ: Wiley, 2010. Print.
}
\seealso{
\code{\link[survey]{withReplicates}} \code{\link[MASS]{polr}} \code{\link[MASS]{glm.nb}}
}

