% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svymisc.R
\name{svypolr}
\alias{svynb}
\alias{svypolr}
\alias{svyrlm}
\title{Wrapper for MASS functions}
\usage{
svypolr(formula, design, subset, ..., scale.weights = FALSE)

svynb(formula, design, subset, ..., scale.weights = FALSE)

svyrlm(formula, design, subset, ..., scale.weights = FALSE)
}
\arguments{
\item{formula}{Model formula}

\item{design}{Survey design from \code{\link[survey]{svrepdesign}}}

\item{subset}{Expression to select a subpopulation}

\item{...}{Other arugments passed to \code{\link[MASS]{polr}} or
\code{\link[MASS]{glm.nb}} or \code{\link[MASS]{rlm}}}

\item{scale.weights}{Indicate whether to rescale weights (defaults to false)}
}
\description{
Wrapper for several functions from \pkg{MASS} for replicate weights
}
\note{
\code{scale.weights} defaults to false but Lumley recommends rescaling weights
}
\seealso{
\code{\link[survey]{withReplicates}} \code{\link[MASS]{polr}} \code{\link[MASS]{glm.nb}}
}

