% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svymisc.R
\name{svyrq}
\alias{svyrq}
\alias{svyrqss}
\title{Wrapper for Quantile Regression}
\usage{
svyrq(formula, design, subset, ..., scale.weights = FALSE)

svyrqss(formula, design, subset, ..., scale.weights = FALSE)
}
\arguments{
\item{formula}{Model formula}

\item{design}{Survey design from \code{\link[survey]{svrepdesign}}}

\item{subset}{Expression to select a subpopulation}

\item{...}{Other arugments passed to \code{\link[quantreg]{rq}} or \code{\link[quantreg]{rqss}}}

\item{scale.weights}{Indicate whether to rescale weights (defaults to false)}
}
\description{
Wrapper for \code{\link[quantreg]{rq}} and \code{\link[quantreg]{rqss}} for replicate weights
}
\references{
Lumley, Thomas. Complex Surveys: A Guide to Analisys Using R.
 Hoboken, NJ: Wiley, 2010. Print.
}
\seealso{
\code{\link[survey]{withReplicates}} \code{\link[quantreg]{rq}} \code{\link[quantreg]{rqss}}
}

