% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svymisc.R
\name{svytrch}
\alias{svytrch}
\alias{svycrch}
\title{Wrapper for Censored and Truncated Response Model}
\usage{
svytrch(formula, design, subset, ..., scale.weights = FALSE)

svycrch(formula, design, subset, ..., scale.weights = FALSE)
}
\arguments{
\item{formula}{Model formula}

\item{design}{Survey design from \code{\link[survey]{svrepdesign}}}

\item{subset}{Expression to select a subpopulation}

\item{...}{Other arugments passed to \code{\link[crch]{trch}} or \code{\link[crch]{crch}}}

\item{scale.weights}{Indicate whether to rescale weights (defaults to false)}
}
\description{
Wrapper for \code{\link[crch]{trch}} and \code{\link[crch]{crch}} for replicate weights
}
\references{
Lumley, Thomas. Complex Surveys: A Guide to Analisys Using R.
 Hoboken, NJ: Wiley, 2010. Print.
}
\seealso{
\code{\link[survey]{withReplicates}} \code{\link[crch]{trch}} \code{\link[crch]{crch}}
}
