% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svymisc.R
\name{svyclm}
\alias{svyclm}
\title{Wrapper for Ordinal Logistic Regression (cumulative link model) for Replicate Weights}
\usage{
svyclm(formula, design, subset, ..., scale.weights = FALSE)
}
\arguments{
\item{formula}{Model formula}

\item{design}{Survey design from \code{\link[survey]{svrepdesign}}}

\item{subset}{Expression to select a subpopulation}

\item{...}{Other arugments passed to \code{\link[ordinal]{clm}}}

\item{scale.weights}{Indicate whether to rescale weights (defaults to false)}
}
\description{
Uses \code{\link[survey]{withReplicates}} and \code{\link[ordinal]{clm}} to generate
coefficients, and standards errors for ordinal logistic regressions
using replicate weights
}
\examples{
library(survey)
data(api)
d <- svydesign(id=~dnum, weights=~pw, data=apiclus1, fpc=~fpc)
dwr <- as.svrepdesign(d, type = "bootstrap", replicates = 100)
mod <- svyclm(stype ~ ell + mobility, dwr)
mod
confint(mod)
}
\references{
Lumley, Thomas. Complex Surveys: A Guide to Analisys Using R.
 Hoboken, NJ: Wiley, 2010. Print.
}
\seealso{
\code{\link[survey]{withReplicates}} \code{\link[ordinal]{clm}}
}
