% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svymisc.R
\name{svyintReg}
\alias{svyintReg}
\title{Wrapper for Interval Regression}
\usage{
svyintReg(formula, design, subset, ..., scale.weights = FALSE)
}
\arguments{
\item{formula}{Model formula}

\item{design}{Survey design from \code{\link[survey]{svrepdesign}}}

\item{subset}{Expression to select a subpopulation}

\item{...}{Other arugments passed to \code{\link[intReg]{intReg}}}

\item{scale.weights}{Indicate whether to rescale weights (defaults to false)}
}
\description{
Wrapper for \code{\link[intReg]{intReg}} for replicate weights
}
\references{
Lumley, Thomas. Complex Surveys: A Guide to Analisys Using R.
 Hoboken, NJ: Wiley, 2010. Print.
}
\seealso{
\code{\link[survey]{withReplicates}} \code{\link[intReg]{intReg}}
}
