% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svymisc.R
\name{svypolr}
\alias{svyglm.nb}
\alias{svypolr}
\alias{svyrlm}
\title{Wrapper for MASS functions}
\usage{
svypolr(formula, design, subset, ...)

svyglm.nb(formula, design, subset, ...)

svyrlm(formula, design, subset, ...)
}
\description{
Wrapper for several functions from \pkg{MASS} for replicate weights
}
\seealso{
\code{\link[survey]{withReplicates}} \code{\link[MASS]{polr}} \code{\link[MASS]{glm.nb}}
}

