% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svymisc.R
\name{svyrq}
\alias{svyrq}
\title{Wrapper for Quantile Regression}
\usage{
svyrq(formula, design, subset, ...)
}
\description{
Wrapper for \code{\link[quantreg]{rq}} for replicate weights
}
\seealso{
\code{\link[survey]{withReplicates}} \code{\link[quantreg]{rq}}
}

