% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svymisc.R
\name{svyrq}
\alias{svyrq}
\title{Wrapper for Quantile Regression}
\usage{
svyrq(formula, design, subset, ..., scale.weights = FALSE)
}
\arguments{
\item{formula}{Model formula}

\item{design}{Survey design from \code{\link[survey]{svrepdesign}}}

\item{subset}{Expression to select a subpopulation}

\item{...}{Other arugments passed to \code{\link[quantreg]{rq}}}

\item{scale.weights}{Indicate whether to rescale weights (defaults to false)}
}
\description{
Wrapper for \code{\link[quantreg]{rq}} for replicate weights
}
\seealso{
\code{\link[survey]{withReplicates}} \code{\link[quantreg]{rq}}
}

