% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svymisc.R
\name{svytruncreg}
\alias{svytruncreg}
\title{Wrapper for Truncated Response Model}
\usage{
svytruncreg(formula, design, subset, ..., scale.weights = FALSE)
}
\arguments{
\item{formula}{Model formula}

\item{design}{Survey design from \code{\link[survey]{svrepdesign}}}

\item{subset}{Expression to select a subpopulation}

\item{...}{Other arugments passed to \code{\link[truncreg]{truncreg}}}

\item{scale.weights}{Indicate whether to rescale weights (defaults to false)}
}
\description{
Wrapper for \code{\link[truncreg]{truncreg}} for replicate weights
}
\note{
\code{scale.weights} defaults to false but Lumley recommends rescaling weights
}

